refactor: VideoContentView to CameraView
